moved mummer to it's own package, small changes in qDna package, wrote new tests for axt and vcf packages, and most importantly a working draft of the genome graph (pairwise alignment) genome graph write to file and read is working

+++ qDna/qDna.go
@@ -24,6 +24,6 @@ type QAdj struct {
 type Location struct {
 	Assembly string
 	Chr      string
-	Start    int
-	End      int
+	Start    int64
+	End      int64
 }
